FIX: Alter "chr" prefix in VCF & beagle map files.
If the chromosome doesn't match, e.g. chr22!=22, then beagle doesn't work. This should be one step, however. In the future, edit to simply remove the "chr" prefix from the VCF first (before phasing) instead of adding it to the map file; it needs to be removed from the VCF post-phasing regardless.