FIX: Remove duplicate sites based on position; replace variant IDs
ADMIXTURE shouldn''t be run on multiallelic sites. Plink also merges based on variant ID instead of position (also due to multiallelic sites), so variants IDs in each dataset need to match. It was easiest to rename all variant IDs to the CHR_POS_REF_ALT format, because there was a significant number of "." variant IDs that don't have rsIDs that will cause plink to have an error. Also made an output directory & README.